Rebase on to main Remove ophan from data collect. This will be added to the launch wizard in a later PR. moved the `editBrazeLayout`, `editIdentityName`, `ophan` and `editOphanLayout` to the launch wizard folder
