+++ libs/newsletter-workflow/src/lib/steps/launchNewsletter/editOphanLayout.ts
@@ -1,22 +1,16 @@
 import type { DraftStorage } from '@newsletters-nx/newsletters-data-client';
 import type { WizardStepLayout } from '@newsletters-nx/state-machine';
-import { executeSkip } from '../../ececuteSkip';
 import { executeModify } from '../../executeModify';
 import { getStringValuesFromRecord } from '../../getValuesFromRecord';
 import { regExPatterns } from '../../regExPatterns';
+import { formSchemas } from '../newsletterData/formSchemas';
 
 const markdownTemplate = `
-# Ophan Campaign Values
+# Modify Ophan Campaign Values
 
 These are tracking fields used by Ophan.
 
 They have been calculated automatically from the name **{{name}}**, but you can change them if you need.
-
-&nbsp;
-
-Campaign Name: **{{campaignName}}**
-
-Campaign Code: **{{campaignCode}}**
 
 `.trim();
 
@@ -25,45 +19,33 @@ const staticMarkdown = markdownTemplate.replace(
 	'of the newsletter',
 );
 
-export const ophanLayout: WizardStepLayout<DraftStorage> = {
+export const editOphanLayout: WizardStepLayout<DraftStorage> = {
 	staticMarkdown,
-	label: 'Ophan',
 	dynamicMarkdown(requestData, responseData) {
 		if (!responseData) {
 			return staticMarkdown;
 		}
 		const [name = 'NAME'] = getStringValuesFromRecord(responseData, ['name']);
-		const [campaignName = 'CAMPAIGNNAME'] = getStringValuesFromRecord(
-			responseData,
-			['campaignName'],
-		);
-		const [campaignCode = 'CAMPAIGNCODE'] = getStringValuesFromRecord(
-			responseData,
-			['campaignCode'],
-		);
-		return markdownTemplate
-			.replace(regExPatterns.name, name)
-			.replace(regExPatterns.campaignName, campaignName)
-			.replace(regExPatterns.campaignCode, campaignCode);
+		return markdownTemplate.replace(regExPatterns.name, name);
 	},
 	buttons: {
 		back: {
 			buttonType: 'RED',
 			label: 'Back',
-			stepToMoveTo: 'signUp',
+			stepToMoveTo: 'signUpEmbed',
 			executeStep: executeModify,
-		},
-		edit: {
-			buttonType: 'GREEN',
-			label: 'Edit',
-			stepToMoveTo: 'editOphan',
 		},
 		next: {
 			buttonType: 'GREEN',
 			label: 'Complete Data Collection',
 			stepToMoveTo: 'completeDataCollection',
+			onBeforeStepChangeValidate: () => {
+				// TO DO - check that ophan values do not already exist in other draft or actual newsletter
+				return undefined;
+			},
+			executeStep: executeModify,
 		},
 	},
-	canSkipTo: true,
-	executeSkip: executeSkip,
+	schema: formSchemas.ophan,
+	parentStepId: 'ophan',
 };
