+++ libs/newsletter-workflow/src/lib/steps/newsletterData/editDraftNewsletterLayout.ts
@@ -1,5 +1,6 @@
 import type { DraftStorage } from '@newsletters-nx/newsletters-data-client';
 import type { WizardStepLayout } from '@newsletters-nx/state-machine';
+import { executeSkip } from '../../ececuteSkip';
 import { executeCreate } from '../../executeCreate';
 import { getDraftFromStorage } from '../../getDraftFromStorage';
 import { formSchemas } from './formSchemas';
@@ -31,4 +32,6 @@ The first step is to enter the name for your newsletter, for example **Down to E
 	schema: formSchemas.startDraftNewsletter,
 	role: 'EDIT_START',
 	getInitialFormData: getDraftFromStorage,
+	canSkipTo: true,
+	executeSkip: executeSkip,
 };
